Implement a new cloud-sync multi-site relation
The new relation is based on the Ceph cloud sync module:
https://docs.ceph.com/en/latest/radosgw/cloud-sync-module/

The cloud sync module leverages the same multi-site replication
framework, but instead of writing to a secondary Ceph cluster, it
writes to an S3 (or S3-compatible) target.

The S3 target connection credentials are obtained through a new
relation with the s3-integrator charm.

This new relation behaves exactly as a secondary multi-site relation
(it implements the same interface), but it is used to configure the
cloud sync module for the primary RGW zone. The secondary Ceph RGW is
related to the Ceph Mons in the same cluster as the primary Ceph RGW.

Change-Id: Ia9b69d2f48e77f73b55a2036ae7845c0cbba8045
func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1176
Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>